requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.137
botocore==1.34.137
distlib==0.3.8
Django==4.2.7
djangorestframework==3.15.1
django-bootstrap-v5==1.0.11
django-filter==24.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
jmespath==1.0.1
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.2
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.25.1

